% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_lists.R
\name{cl_codes}
\alias{cl_codes}
\title{SDTM terminology codelists' codes}
\usage{
cl_codes(extensible = c("any", "yes", "no"))
}
\arguments{
\item{extensible}{Should the returned value be restricted to codelists that
are extensible (\code{"yes"}), not extensible (\code{"no"}), or both
(\code{"any"}).}
}
\value{
A character vector of codelists' codes.
}
\description{
\code{\link[=cl_codes]{cl_codes()}} returns the set of possible codelists' codes.
}
\examples{
# `cl_codes()` is the same as `cl_codes(extensible = "any")`
cl_codes()

cl_codes(extensible = "yes")

cl_codes(extensible = "no")

}
