% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_lists.R
\name{cl_definitions}
\alias{cl_definitions}
\title{SDTM terminology codelists' definitions}
\usage{
cl_definitions(extensible = c("any", "yes", "no"))
}
\arguments{
\item{extensible}{Should the returned value be restricted to codelists that
are extensible (\code{"yes"}), not extensible (\code{"no"}), or both
(\code{"any"}).}
}
\value{
A character vector of codelists' definitions.
}
\description{
\code{\link[=cl_definitions]{cl_definitions()}} returns the set of possible codelists' definitions.
}
\examples{
# `cl_definitions()` is the same as `cl_definitions(extensible = "any")`
cl_definitions()

cl_definitions(extensible = "yes")

cl_definitions(extensible = "no")

}
