% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_lists.R
\name{cl_names}
\alias{cl_names}
\title{SDTM terminology codelists' names}
\usage{
cl_names(extensible = c("any", "yes", "no"))
}
\arguments{
\item{extensible}{Should the returned value be restricted to codelists that
are extensible (\code{"yes"}), not extensible (\code{"no"}), or both
(\code{"any"}).}
}
\value{
A character vector of codelists' names.
}
\description{
\code{\link[=cl_names]{cl_names()}} returns the set of possible codelists' names.
}
\examples{
# `cl_names()` is the same as `cl_names(extensible = "any")`
cl_names()

cl_names(extensible = "yes")

cl_names(extensible = "no")

}
