% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_lists.R
\name{cl_nci_terms}
\alias{cl_nci_terms}
\title{SDTM terminology codelists' NCIt preferred names.}
\usage{
cl_nci_terms(extensible = c("any", "yes", "no"))
}
\arguments{
\item{extensible}{Should the returned value be restricted to codelists that
are extensible (\code{"yes"}), not extensible (\code{"no"}), or both
(\code{"any"}).}
}
\value{
A character vector of codelists' NCIt preferred names.
}
\description{
\code{\link[=cl_nci_terms]{cl_nci_terms()}} returns the set of possible codelists' NCIt preferred names.
}
\examples{
# `cl_nci_terms()` is the same as `cl_nci_terms(extensible = "any")`
cl_nci_terms()

cl_nci_terms(extensible = "yes")

cl_nci_terms(extensible = "no")

}
