% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_lists.R
\name{cl_synonyms}
\alias{cl_synonyms}
\title{SDTM terminology codelists' synonyms}
\usage{
cl_synonyms(extensible = c("any", "yes", "no"))
}
\arguments{
\item{extensible}{Should the returned value be restricted to codelists that
are extensible (\code{"yes"}), not extensible (\code{"no"}), or both
(\code{"any"}).}
}
\value{
A character vector of codelists' synonyms.
}
\description{
\code{\link[=cl_synonyms]{cl_synonyms()}} returns the set of possible codelists' synonyms.
}
\examples{
# `cl_synonyms()` is the same as `cl_synonyms(extensible = "any")`
cl_synonyms()

cl_synonyms(extensible = "yes")

cl_synonyms(extensible = "no")

}
