% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_lists.R
\name{cl_terms}
\alias{cl_terms}
\title{SDTM terminology codelists' terms}
\usage{
cl_terms(extensible = c("any", "yes", "no"))
}
\arguments{
\item{extensible}{Should the returned value be restricted to codelists that
are extensible (\code{"yes"}), not extensible (\code{"no"}), or both
(\code{"any"}).}
}
\value{
A character vector of codelists' terms.
}
\description{
\code{\link[=cl_terms]{cl_terms()}} returns the set of possible codelists' terms.
}
\examples{
# `cl_terms()` is the same as `cl_terms(extensible = "any")`
cl_terms()

cl_terms(extensible = "yes")

cl_terms(extensible = "no")

}
