% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct-release.R
\name{ct_release}
\alias{ct_release}
\title{CDISC SDTM Controlled Terminology Release Date}
\usage{
ct_release()
}
\value{
A \link[base:Dates]{Date} object.
}
\description{
SDTM terminology release date.
}
\examples{
ct_release()

}
