% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ct.R
\name{is_term}
\alias{is_term}
\alias{is_code}
\title{Is a code or term in a codelist?}
\usage{
is_term(term, codelist)

is_code(code, codelist)
}
\arguments{
\item{term}{A character vector of terms (e.g. \code{"TENMW102"}).}

\item{codelist}{A character vector of codes referring to SDTM terminology
codelists (e.g. \code{"C141657"}).}

\item{code}{A character vector of C-codes (e.g. \code{"C174106"}).}
}
\value{
A \code{logical} vector, where \code{TRUE} indicates that \code{term} is a
controlled terminology term within the \code{codelist} provided, and \code{FALSE},
otherwise.
}
\description{
\code{\link[=is_code]{is_code()}} and \code{\link[=is_term]{is_term()}} evaluate whether query codes or terms are part of
the SDTM controlled terminology within the context of a codelist.
}
\examples{
is_term(term = "TENMW101", codelist = "C141657")

# Note that the codelist term itself (`"TENMW1TC"`) is not part of its own
# codelist (`"C141657"`).
is_term(term = "TENMW1TC", codelist = "C141657")

# `is_term()` is vectorized over `term` and `codelist`
is_term(term = c("TENMW101", "TENMW104"), codelist = "C141657")

is_code(code = "C174106", codelist = "C141657")

# Note that the codelist code itself (`"C141657"`) is not part of its own
# codelist (`"C141657"`).
is_code(code = "C141657", codelist = "C141657")

# `is_code()` is vectorized over `code` and `codelist`
is_code(code = c("C174106", "C141701"), codelist = "C141657")

}
