% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-version.R
\name{pkg_version}
\alias{pkg_version}
\title{Get package version}
\usage{
pkg_version()
}
\value{
Package version as an object of class
\link[base:numeric_version]{package_version}.
}
\description{
\code{\link[=pkg_name]{pkg_name()}} returns package version
}
\keyword{internal}
