% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.terminology-package.R
\docType{package}
\name{sdtm.terminology-package}
\alias{sdtm.terminology}
\alias{sdtm.terminology-package}
\title{sdtm.terminology: CDISC SDTM Controlled Terminology}
\description{
Clinical Data Interchange Standards Consortium (CDISC) Standard Data Tabulation Model (SDTM) Controlled Terminology, 2025-03-25. Source: \url{https://evs.nci.nih.gov/ftp1/CDISC/SDTM/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/patterninstitute/sdtm.terminology}
  \item \url{https://www.pattern.institute/sdtm.terminology/}
  \item Report bugs at \url{https://github.com/patterninstitute/sdtm.terminology/issues}
}

}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Other contributors:
\itemize{
  \item Pattern Institute [copyright holder, funder]
}

}
\keyword{internal}
