% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdtm_code_lists}
\alias{sdtm_code_lists}
\title{SDTM code lists}
\format{
A \link[tibble:tibble-package]{tibble} providing a data set of controlled
terms for SDTM code lists.

\describe{
\item{\code{code}}{A C-code with a leading character C followed by a sequence of
digits used for uniquely identifying each concept in NCI Thesaurus (NCIt),
including all CDISC concepts.}
\item{\code{term}}{This is the specific value expected for submissions.}
\item{\code{is_extensible}}{Defines if controlled terms may be added to the
codelist. New terms may be added to existing codelist values when the
codelist has value \code{TRUE} as long as the terms are not duplicates or synonyms
of existing terms. The expectation is that sponsors will use the published
controlled terminology as a standard baseline and codelists defined as
extensible may have terms added by the sponsor internally.  For codelists
with \code{FALSE} sponsor organizations may not add to the published lists.
Suggestions for new terms may be requested through the New Term Request
Page.}
\item{\code{cl_name}}{Contains the descriptive name of the codelist.}
\item{\code{synonym}}{This identifies the applicable synonyms for the CDISC
submission value indicated in column \code{term}. These terms should not be
submitted, but are included for collection and mapping purposes. Note: this
is especially important in instances where a test name or parameter test name
contains a corresponding test code or parameter test code.}
\item{\code{definition}}{This identifies the value of an \code{ALT_DEFINITION} property
of a CDISC source for a particular concept. In many cases an existing NCI
definition has been used. The source for a definition is noted in parentheses
(e.g. NCI, CDISC glossary).}
\item{\code{nci_term}}{This identifies the value of a \code{PREFERRED_NAME} property for
a concept in NCIt. This column designates the human readable, fully specified
preferred term corresponding to the NCI C-code, and is especially helpful for
searching NCIt to get the entire concept with links to all instances of the
term.}
}
}
\source{
\url{https://evs.nci.nih.gov/ftp1/CDISC/SDTM/}
}
\usage{
sdtm_code_lists
}
\description{
\link{sdtm_code_lists} is a subset of the data provided in provided in
\link{sdtm_terminology}. It contains only the code lists' terms from CDISC SDTM
Terminology.
}
\examples{
sdtm_code_lists

}
\seealso{
\itemize{
\item \link{sdtm_terminology} for all terms included in CDISC SDTM Terminology.
\item \link{sdtm_code_terms} for other terms (excluding code lists).
}
}
\keyword{datasets}
