% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdtm_code_terms}
\alias{sdtm_code_terms}
\title{SDTM controlled terms (exclusive of code lists)}
\format{
A \link[tibble:tibble-package]{tibble} providing a data set of controlled
terms, excluding terms for code lists.

\describe{
\item{\code{cl_code}}{An NCIt C-code assigned to the parent codelist name. This
code is repeated for each concept (aka permissible value) belonging to a
codelist.}
\item{\code{code}}{A C-code with a leading character C followed by a sequence of
digits used for uniquely identifying each concept in NCI Thesaurus (NCIt),
including all CDISC concepts.}
\item{\code{term}}{This is the specific value expected for submissions.}
\item{\code{synonym}}{This identifies the applicable synonyms for the CDISC
submission value indicated in column \code{term}. These terms should not be
submitted, but are included for collection and mapping purposes. Note: this
is especially important in instances where a test name or parameter test name
contains a corresponding test code or parameter test code.}
\item{\code{definition}}{This identifies the value of an \code{ALT_DEFINITION} property
of a CDISC source for a particular concept. In many cases an existing NCI
definition has been used. The source for a definition is noted in parentheses
(e.g. NCI, CDISC glossary).}
\item{\code{nci_term}}{This identifies the value of a \code{PREFERRED_NAME} property for
a concept in NCIt. This column designates the human readable, fully specified
preferred term corresponding to the NCI C-code, and is especially helpful for
searching NCIt to get the entire concept with links to all instances of the
term.}
}
}
\source{
\url{https://evs.nci.nih.gov/ftp1/CDISC/SDTM/}
}
\usage{
sdtm_code_terms
}
\description{
\link{sdtm_code_terms} is a subset of the data provided in provided in
\link{sdtm_terminology}. It contains controlled terms from CDISC SDTM
Terminology, excluding the terms for code lists. Use \link{sdtm_code_lists}
instead for terms for code lists only.
}
\examples{
sdtm_code_terms

}
\seealso{
\itemize{
\item \link{sdtm_terminology} for all terms included in CDISC SDTM Terminology.
\item \link{sdtm_code_lists} for terms corresponding to code lists.
}
}
\keyword{datasets}
