% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{sdtm_ct_version}
\alias{sdtm_ct_version}
\title{SDTM controlled terminology version}
\usage{
sdtm_ct_version()
}
\value{
A list of three strings:
\describe{
\item{\code{quarter}}{Quarter.}
\item{\code{release}}{Release date.}
\item{\code{SDTMIG}}{Study Data Tabulation Model Implementation Guide version that
is supported by this Controlled Terminology.}
}
}
\description{
\code{\link[=sdtm_ct_version]{sdtm_ct_version()}} returns the version of the SDTM controlled terminology
data that this package is built upon.
}
\examples{
sdtm_ct_version()

}
