% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdtm_terminology}
\alias{sdtm_terminology}
\title{SDTM Terminology}
\format{
A \link[tibble:tibble-package]{tibble} providing a data set of controlled
terms: either terms for code lists or individual terms representing CDISC
concepts subjected to controlled terminology. These two cases can be
distinguished by the variable \code{cl_code}: the parent code list code. A term
representing a code list won't have a parent code list code, i.e. \code{cl_code}
is \code{NA}.

\describe{
\item{\code{code}}{A C-code with a leading character C followed by a sequence of
digits used for uniquely identifying each concept in NCI Thesaurus (NCIt),
including all CDISC concepts.}
\item{\code{term}}{This is the specific value expected for submissions.}
\item{\code{cl_code}}{An NCIt C-code assigned to the parent codelist name. This
code is repeated for each concept (aka permissible value) belonging to a
codelist.}
\item{\code{is_extensible}}{Defines if controlled terms may be added to the
codelist. New terms may be added to existing codelist values when the
codelist has value \code{TRUE} as long as the terms are not duplicates or synonyms
of existing terms. The expectation is that sponsors will use the published
controlled terminology as a standard baseline and codelists defined as
extensible may have terms added by the sponsor internally.  For codelists
with \code{FALSE} sponsor organizations may not add to the published lists.
Suggestions for new terms may be requested through the New Term Request
Page.}
\item{\code{cl_name}}{}
\item{\code{synonym}}{This identifies the applicable synonyms for the CDISC
submission value indicated in column \code{term}. These terms should not be
submitted, but are included for collection and mapping purposes. Note: this
is especially important in instances where a test name or parameter test name
contains a corresponding test code or parameter test code.}
\item{\code{definition}}{This identifies the value of an \code{ALT_DEFINITION} property
of a CDISC source for a particular concept. In many cases an existing NCI
definition has been used. The source for a definition is noted in parentheses
(e.g. NCI, CDISC glossary).}
\item{\code{nci_term}}{This identifies the value of a \code{PREFERRED_NAME} property for
a concept in NCIt. This column designates the human readable, fully specified
preferred term corresponding to the NCI C-code, and is especially helpful for
searching NCIt to get the entire concept with links to all instances of the
term.}
}
}
\source{
\url{https://evs.nci.nih.gov/ftp1/CDISC/SDTM/}
}
\usage{
sdtm_terminology
}
\description{
\link{sdtm_terminology} provides CDISC SDTM Terminology data in
\link[tibble:tibble-package]{tibble} format. It follows CDISC original structure
with a few modifications, see Details section.
}
\details{
\itemize{
\item Column names have been shortened and changed to follow snake case style.
\item The original variable \verb{Codelist Extensible (Yes/No)} whose type is text,
is here renamed to \code{is_extensible} and made a \code{logical} vector with values
recoded from \code{"No"} and \code{"Yes"} to \code{FALSE} and \code{TRUE}, respectively.
}
}
\examples{
sdtm_terminology

}
\seealso{
\itemize{
\item \link{sdtm_code_lists} for terms referring to code lists only.
\item \link{sdtm_code_terms} for other terms (excluding code lists).
}
}
\keyword{datasets}
