% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.terminology-package.R
\docType{package}
\name{sdtm.terminology-package}
\alias{sdtm.terminology}
\alias{sdtm.terminology-package}
\title{sdtm.terminology: CDISC SDTM Controlled Terminology}
\description{
CDISC SDTM Controlled Terminology \url{https://evs.nci.nih.gov/ftp1/CDISC/SDTM/}, Q4 2023, 2023-12-15.
}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Other contributors:
\itemize{
  \item Pattern Institute [copyright holder, funder]
}

}
\keyword{internal}
